refactor: [EL-5114] refactor the code according to comments
