+++ src/[fsd]/features/credentials/lib/hooks/useCredentialValidation.hooks.js
@@ -57,6 +57,53 @@ export const useCredentialValidation = () => {
     [testConnection],
   );
 
+  const validateProjectBatch = useCallback(
+    async (projectId, creds) => {
+      const items = creds.map(credential => ({
+        id: String(credential.originalId || credential.id || credential.uuid),
+        type: credential.type,
+        data: credential.data || credential.settings || {},
+      }));
+
+      try {
+        const result = await batchTestConnection({ projectId: Number(projectId), items });
+        const updates = {};
+
+        if (result.data) {
+          const msgUpdates = {};
+          result.data.forEach(item => {
+            if (item.unsupported) {
+              updates[item.id] = 'unsupported';
+            } else {
+              updates[item.id] = item.success ? 'valid' : 'invalid';
+              if (!item.success && item.message) {
+                msgUpdates[item.id] = item.message;
+              }
+            }
+          });
+          if (Object.keys(msgUpdates).length > 0) {
+            setMessages(prev => ({ ...prev, ...msgUpdates }));
+          }
+        } else {
+          creds.forEach(credential => {
+            const key = String(credential.originalId || credential.id || credential.uuid);
+            updates[key] = 'invalid';
+          });
+        }
+
+        setStatuses(prev => ({ ...prev, ...updates }));
+      } catch {
+        const updates = {};
+        creds.forEach(credential => {
+          const key = String(credential.originalId || credential.id || credential.uuid);
+          updates[key] = 'invalid';
+        });
+        setStatuses(prev => ({ ...prev, ...updates }));
+      }
+    },
+    [batchTestConnection],
+  );
+
   /**
    * Validate multiple credentials in a single batch request per project.
    * Credentials already in a non-idle status are skipped.
@@ -92,53 +139,10 @@ export const useCredentialValidation = () => {
       });
 
       await Promise.all(
-        Object.entries(byProject).map(async ([projectId, creds]) => {
-          const items = creds.map(credential => ({
-            id: String(credential.originalId || credential.id || credential.uuid),
-            type: credential.type,
-            data: credential.data || credential.settings || {},
-          }));
-
-          try {
-            const result = await batchTestConnection({ projectId: Number(projectId), items });
-            const updates = {};
-
-            if (result.data) {
-              const msgUpdates = {};
-              result.data.forEach(item => {
-                if (item.unsupported) {
-                  updates[item.id] = 'unsupported';
-                } else {
-                  updates[item.id] = item.success ? 'valid' : 'invalid';
-                  if (!item.success && item.message) {
-                    msgUpdates[item.id] = item.message;
-                  }
-                }
-              });
-              if (Object.keys(msgUpdates).length > 0) {
-                setMessages(prev => ({ ...prev, ...msgUpdates }));
-              }
-            } else {
-              // Entire batch request failed — mark all as valid
-              creds.forEach(credential => {
-                const key = String(credential.originalId || credential.id || credential.uuid);
-                updates[key] = 'valid';
-              });
-            }
-
-            setStatuses(prev => ({ ...prev, ...updates }));
-          } catch {
-            const updates = {};
-            creds.forEach(credential => {
-              const key = String(credential.originalId || credential.id || credential.uuid);
-              updates[key] = 'valid';
-            });
-            setStatuses(prev => ({ ...prev, ...updates }));
-          }
-        }),
+        Object.entries(byProject).map(([projectId, creds]) => validateProjectBatch(projectId, creds)),
       );
     },
-    [batchTestConnection],
+    [validateProjectBatch],
   );
 
   const getCredentialStatus = useCallback(credentialId => statuses[credentialId] ?? 'idle', [statuses]);
