+++ src/[fsd]/features/credentials/ui/index.js
@@ -1,2 +1,4 @@
 export { default as CredentialsTabBar } from './credentials-tab-bar/CredentialsTabBar';
 export { default as CredentialsControls } from './credentials-tab-bar/CredentialsControls';
+export { default as CredentialOptionLabel } from './credential-option-label/CredentialOptionLabel';
+export { default as CredentialsSelect } from './credentials-select/CredentialsSelect';
