+++ src/[fsd]/features/toolkits/indexes/ui/IndexDetails/IndexScheduleModal.jsx
@@ -6,11 +6,11 @@ import { useSelector } from 'react-redux';
 
 import { Box, Button, GlobalStyles, Typography } from '@mui/material';
 
+import { CredentialsSelect } from '@/[fsd]/features/credentials/ui';
 import { IndexCronDefault } from '@/[fsd]/features/toolkits/indexes/lib/constants/indexDetails.constants';
 import { validateCronExpression } from '@/[fsd]/features/toolkits/indexes/lib/helpers/indexSchedule.helpers.js';
 import { Checkbox, Modal } from '@/[fsd]/shared/ui';
 import InfoTooltip from '@/[fsd]/shared/ui/tooltip/InfoTooltip';
-import CredentialsSelect from '@/components/CredentialsSelect';
 import FormInput from '@/components/FormInput';
 import { useSelectedProject } from '@/hooks/useSelectedProject';
